requirements.txt
Django==1.9.4
psycopg2==2.6.1
pytz==2016.3
django-tastypie==0.13.3